// src/features/doctors/types/doctor.types.ts
export type DoctorRaw = {
  ID: number | null;
  Obj_Name: string | null;
  Active: boolean | null;
  IDNo: string | null;
  Address: string | null;
  Phone: string | null;
  Mobile: string | null;
  Email: string | null;
  Gender_Name: string | null;
  Gender_ID: number | null;
  Branch_Name: string | null;
  Branch_ID: number | null;
  SocialID: string | null;
  Marital_Name: string | null;
  Marital_Status: number | null;
  Millitary_Name: string | null;
  Millitary_ID: number | null;
  BirthDate: string | null;
  Country_Name: string | null;
  Country_ID: number | null;
  State_Name: string | null;
  State_ID: number | null;
  City_Name: string | null;
  City_ID: number | null;
  Notes: string | null;
  Specialty_Name: string | null;
  Specialty_ID: number | null;
};
export type Doctor = {
  id: number;
  name: string;
  active: boolean;
  nationalId: string;
  address: string;
  phone: string;
  mobile: string;
  email: string;
  genderName: string;
  genderId: number | null;
  branchName: string;
  branchId: number | null;
  socialId: string;
  maritalStatusName: string;
  maritalStatusId: number | null;
  militaryStatusName: string;
  militaryStatusId: number | null;
  birthDate: string;
  countryName: string;
  countryId: number | null;
  stateName: string;
  stateId: number | null;
  cityName: string;
  cityId: number | null;
  notes: string;
  specialtyName: string;
  specialtyId: number;
};